Use a sqlite database with Peewee ORM to store records.
Had to make a repeating string field entry since one wasn't already available. Opted for this method instead of normalising because there was no reason to add the complexity.
DogAct no longer stores any non-primitive values, so it now does lookups to get the members when necessary. Slightly slower but not noticeable.
Database is stored in a place that can be made a mounted volume for deployment persistence.