CLI config params, custom bot class, bugs
Fixed interactions causing an error on the user side by deferring their interaction.

Parameters can now be provided to the command line when executing the bot, which overload the config if that's set to be used also.
Because the config is now a parameter, the database file needs to be loaded in dynamically. To achieve this, I swapped to a database proxy which is set on app start.
Stripped owner ids from default config since that's a security concern.

Part of achieving this change was to move the bot events into a custom bot extension class, which hooks into the `disnake.Bot` API automatically to achieve the same outcome.
Doing this will make the application testable also.

In order to accommodate for this new way of starting the app, the command had to be updated within the docker-compose.yml file too.